[refactor] 파일 위치 이동, 컴포넌트 내부 폴더 생성
