+++ src/app/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from "next";
 import { Geist, Geist_Mono } from "next/font/google";
 import "@/styles/index.scss";
-import Header from "@/components/Header/Header";
+import Header from "@/components/commons/Header/Header";
 
 const geistSans = Geist({
   variable: "--font-geist-sans",
